Separate ENS library logic from Argent-specific
