+++ contracts/ens/ENSReverseRegistrar.sol
@@ -16,15 +16,11 @@
 pragma solidity ^0.5.4;
 
 /**
- * @dev Interface for an ENS Mananger.
+ * @dev ENS Reverse Registrar interface.
  */
-interface IENSManager {
-    event RootnodeOwnerChange(bytes32 indexed _rootnode, address indexed _newOwner);
-    event ENSResolverChanged(address addr);
-    event Registered(address indexed _owner, string _ens);
-    event Unregistered(string _ens);
-
-    function changeRootnodeOwner(address _newOwner) external;
-    function register(string calldata _label, address _owner) external;
-    function isAvailable(bytes32 _subnode) external view returns(bool);
+contract ENSReverseRegistrar {
+    function claim(address _owner) public returns (bytes32);
+    function claimWithResolver(address _owner, address _resolver) public returns (bytes32);
+    function setName(string memory _name) public returns (bytes32);
+    function node(address _addr) public pure returns (bytes32);
 }
