+++ deployment/2_deploy_contracts.js
@@ -2,7 +2,7 @@ const BaseWallet = require('../build/BaseWallet');
 const ModuleRegistry = require('../build/ModuleRegistry');
 const CompoundRegistry = require('../build/CompoundRegistry');
 const MultiSig = require('../build/MultiSigWallet');
-const ENS = require('../build/TestENSRegistry');
+const ENS = require('../build/ENSRegistryWithFallback');
 const ENSManager = require('../build/ArgentENSManager');
 const ENSResolver = require('../build/ArgentENSResolver');
 const WalletFactory = require('../build/WalletFactory');
