+++ deployment/999_benchmark.js
@@ -5,7 +5,7 @@ const Proxy = require('../build/Proxy');
 const ModuleRegistry = require('../build/ModuleRegistry');
 const DappRegistry = require('../build/DappRegistry');
 const MultiSig = require('../build/MultiSigWallet');
-const ENS = require('../build/TestENSRegistry');
+const ENS = require('../build/ENSRegistryWithFallback');
 const ENSManager = require('../build/ArgentENSManager');
 const ENSResolver = require('../build/ArgentENSResolver');
 const WalletFactory = require('../build/WalletFactory');
