+++ test/factory.js
@@ -1,10 +1,11 @@
 const Wallet = require("../build/BaseWallet");
 const Module = require("../build/BaseModule");
 const ModuleRegistry = require("../build/ModuleRegistry");
-const ENS = require('../build/TestENSRegistry');
+const ENSRegistry = require('../build/ENSRegistry');
+const ENSRegistryWithFallback = require('../build/ENSRegistryWithFallback');
 const ENSManager = require('../build/ArgentENSManager');
 const ENSResolver = require('../build/ArgentENSResolver');
-const ENSReverseRegistrar = require('../build/TestReverseRegistrar');
+const ENSReverseRegistrar = require('../build/ReverseRegistrar');
 const Factory = require('../build/WalletFactory');
 
 const TestManager = require("../utils/test-manager");
@@ -35,7 +36,8 @@ describe("Test Wallet Factory", function () {
 
     before(async () => {
         deployer = manager.newDeployer();
-        ensRegistry = await deployer.deploy(ENS);
+        const ensRegistryWithoutFallback = await deployer.deploy(ENSRegistry);
+        ensRegistry = await deployer.deploy(ENSRegistryWithFallback, {}, ensRegistryWithoutFallback.contractAddress);
         ensResolver = await deployer.deploy(ENSResolver);
         ensReverse = await deployer.deploy(ENSReverseRegistrar, {}, ensRegistry.contractAddress, ensResolver.contractAddress);
         ensManager = await deployer.deploy(ENSManager, {}, subnameWallet + '.' + root, walletNode, ensRegistry.contractAddress, ensResolver.contractAddress);
